requirements.txt
Flask==3.1.0
qrcode[pil]==8.0
Pillow==11.1.0
segno==1.6.1
pypdf==5.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
groq==0.25.0
